video_ingestion_service/requirements.txt
Flask==3.1.0
google-cloud-storage==2.19.0
gunicorn==23.0.0
Flask-Cors==5.0.0
opencv-python==4.11.0.86